more test values
