+++ __tests__/Utils-test.js
@@ -4,7 +4,7 @@
 
 import 'react-native';
 import React from 'react';
-import {shuffle, createArrayOfNumbers, mergeTwoArrays, checkWhetherAllCardsMatched} from '../src/Utils/SupportFun'
+import {shuffle, createArrayOfNumbers, mergeTwoArrays, checkWhetherAllCardsMatched, arrayToDataArray} from '../src/Utils/SupportFun'
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
@@ -71,14 +71,6 @@ it('mergeTwoArrays test number of elements after merge', async() => {
     expect(mergeTwoArrays(arrayOne, arrayTwo)).toHaveLength(12)
 });
 
-/*
-checkWhetherAllCardsMatched
- let data = {
-          value: number,
-          id:index,
-          flipped:false,
-        }
-*/
 
 it('checkWhetherAllCardsMatched test array false', async() => {
     let sampleArray =  [{
@@ -124,4 +116,32 @@ it('checkWhetherAllCardsMatched test array true', async() => {
       }]
     
     expect(checkWhetherAllCardsMatched(sampleArray)).toBeTruthy();
+});
+
+it('arrayToDataArray test array length', async() => {
+    let sampleArray =  [1,2,4]
+    expect(arrayToDataArray(sampleArray)).toHaveLength(3);
+});
+
+it('arrayToDataArray test array length', async() => {
+    let sampleArray =  [1,2,4]
+    expect(arrayToDataArray(sampleArray)).toHaveLength(3);
+});
+
+it('arrayToDataArray test array values', async() => {
+    let sampleArray =  [1,2,4]
+    let sampleDataArray = arrayToDataArray(sampleArray)
+
+
+    expect(sampleDataArray[0].value == 1).toBeTruthy();
+    expect(sampleDataArray[0].id == 0).toBeTruthy();
+    expect(sampleDataArray[0].flipped).toBeFalsy();
+
+    expect(sampleDataArray[1].value == 2).toBeTruthy();
+    expect(sampleDataArray[1].id == 1).toBeTruthy();
+    expect(sampleDataArray[1].flipped).toBeFalsy();
+
+    expect(sampleDataArray[2].value == 4).toBeTruthy();
+    expect(sampleDataArray[2].id == 2).toBeTruthy();
+    expect(sampleDataArray[2].flipped).toBeFalsy();
 });
